runtime: builtins: move to new bank submodule
